requirements.txt
Flask
pandas
openpyxl
fpdf
Pillow
python-docx
python-magic-bin ; sys_platform == "win32"
python-magic ; sys_platform != "win32"